requirements.txt
pyaudio==0.2.13
openai==0.27.2
google-cloud-speech==2.18.0
pygame==2.3.0
google-cloud-texttospeech==2.14.1
keyboard==0.13.5
pyperclip==1.8.2
